Move font weight and style out of the font face
When the weight and styles are defined in the component that uses a
specific font-family, having the properties in the component makes the
styles degrade in a more graceful way.

For example, if the font is blocked or doesn't download, bold styles
still show bold text. This would not happen if the weight was only
defined in the font-face declaration.