Fix font file double download and page change flash
By default Gatsby unmounts the whole page when navigating between
pages. This resulted in also mounting the global styles and triggering
a font download for every page change. Wrapping the Gatsby root
element with a Root component that is not unmounted between page
changes avoided this problem.

To fix the initial page load downloading the font files twice, the
font faces had to be put into a CSS file so the Webpack runtime didn't
download them in addition to the browser.