Use font-display: fallback
On a slow network, we want to show the fallback font so the user
doesn't have to wait for the font to download to see the content.

However, on a fast network or with the font file already cached, we
don't want to flash the fallback font if the actual font would render
quickly anyways.

Using the `fallback` value gives us a short block period with
invisible text before rendering the fallback font.