Reviewed concepts
This is a good occasion to establish a rule regarding the use of Flex application concepts. I suggest the following:

Flex Template for Web and Console are brand names. They are used as per the grammar regarding proper nouns: without the definitive article ("customize Flex Template for Web"; "applications located in Console", unless they are used as attributive modifiers to common nouns ("in the FTW documentation").

Flex has (in the future) two different API's. One is called Marketplace, the other is called Admin. "API" is a common noun, to which the proper noun "Marketplace" or "Admin" is an attributive modifier. Thus, the API names are used with the definitive article: "Features powered by the Marketplace API"; The Admin API can be used to..."