Time-based plan required
The wording that says that Flex API defaults to day-based plan *which is always available* is very confusing. The actual effect that the users will see if they don't provide time-based plan for time-based listing is that there is no availability. Although the "day-based plan that is always available" is internally correct, it's not the behaviour the users will see.