Remove "contact support if you have forgotten password".
The Forgot password link has been added to Console. It's easy to find for a user who tries to log in to Console but has forgotten the password. I don't think we need to mention about that in Docs. People know how to use that feature to request new password.